Move code related to the CPU in `cpu.rs`
